Improved byte packing/unpacking and implemented DhcpPacket.str()
This change includes many updates to the byte packing/unpacking
routines that convert from raw data off the wire to byte arrays
and vice versa.

This change also includes an implementation of DhcpPacket.str()
which returns a formatted human-readable representation of the
DHCP packet and all fields and options.

Beginning of DhcpOffer class.